+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Current Master
 
+- Update gasLimit params with every new block seen.
+
 ## 3.7.7 2017-6-8
 
 - Fix bug where metamask would show old data after computer being asleep or disconnected from the internet.